Test fixture file renames, style, comments.
