+++ test/protocols/bitcoind/bitcoind_rpc.cpp
@@ -17,7 +17,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 #include "../../test.hpp"
-#include "bitcoind.hpp"
+#include "bitcoind_setup_fixture.hpp"
 
 BOOST_FIXTURE_TEST_SUITE(bitcoind_tests, bitcoind_ten_block_setup_fixture)
