+++ test/protocols/bitcoind/bitcoind_setup_fixture.hpp
