+++ test/protocols/electrum/electrum_subscribe.cpp
@@ -17,7 +17,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 #include "../../test.hpp"
-#include "electrum.hpp"
+#include "electrum_setup_fixture.hpp"
 
 using namespace system;
 static const code not_found{ server::error::not_found };
@@ -109,7 +109,7 @@ BOOST_AUTO_TEST_CASE(electrum__blockchain_address_subscribe__repeat_call__idempo
 {
     BOOST_REQUIRE(handshake(electrum::version::v1_0));
 
-    // This validates curosr/midstate consistency.
+    // This validates cursor/midstate consistency.
     BOOST_REQUIRE(query_.set(test::bogus_block10, database::context{ 0, 10, 0 }, false, false));
     BOOST_REQUIRE(query_.set(test::bogus_block11, database::context{ 0, 11, 0 }, false, false));
     BOOST_REQUIRE(query_.set(test::bogus_block12, database::context{ 0, 12, 0 }, false, false));
@@ -139,7 +139,7 @@ BOOST_AUTO_TEST_CASE(electrum__blockchain_address_subscribe__progressive__expect
 {
     BOOST_REQUIRE(handshake(electrum::version::v1_0));
 
-    // This validates curosr/midstate consistency.
+    // This validates cursor/midstate consistency.
     BOOST_REQUIRE(query_.set(test::bogus_block10, database::context{ 0, 10, 0 }, false, false));
     BOOST_REQUIRE(query_.set(test::bogus_block11, database::context{ 0, 11, 0 }, false, false));
     BOOST_REQUIRE(query_.set(test::bogus_block12, database::context{ 0, 12, 0 }, false, false));
@@ -192,7 +192,7 @@ BOOST_AUTO_TEST_CASE(electrum__blockchain_address_subscribe__progressive_notify_
 {
     BOOST_REQUIRE(handshake(electrum::version::v1_0));
 
-    // This validates curosr/midstate consistency.
+    // This validates cursor/midstate consistency.
     BOOST_REQUIRE(query_.set(test::bogus_block10, database::context{ 0, 10, 0 }, false, false));
     BOOST_REQUIRE(query_.set(test::bogus_block11, database::context{ 0, 11, 0 }, false, false));
     BOOST_REQUIRE(query_.set(test::bogus_block12, database::context{ 0, 12, 0 }, false, false));
@@ -229,14 +229,14 @@ BOOST_AUTO_TEST_CASE(electrum__blockchain_address_subscribe__progressive_notify_
     const auto notification1 = receive();
     REQUIRE_NO_THROW_TRUE(notification1.at("method").is_string());
     REQUIRE_NO_THROW_TRUE(notification1.at("params").is_array());
-    BOOST_CHECK_EQUAL(notification1.at("method").as_string(), "blockchain.address.subscribe");
+    BOOST_REQUIRE_EQUAL(notification1.at("method").as_string(), "blockchain.address.subscribe");
 
     const auto& params1 = notification1.at("params").as_array();
     BOOST_REQUIRE_EQUAL(params1.size(), 2u);
-    BOOST_CHECK(params1.at(0).is_string());
-    BOOST_CHECK(params1.at(1).is_string());
-    BOOST_CHECK_EQUAL(params1.at(0).as_string(), found_scripthash);
-    BOOST_CHECK_EQUAL(params1.at(1).as_string(), expected_confirm11);
+    BOOST_REQUIRE(params1.at(0).is_string());
+    BOOST_REQUIRE(params1.at(1).is_string());
+    BOOST_REQUIRE_EQUAL(params1.at(0).as_string(), found_scripthash);
+    BOOST_REQUIRE_EQUAL(params1.at(1).as_string(), expected_confirm11);
 
     // Confirming block 12 only makes block 12 confirmed.
     BOOST_REQUIRE(query_.push_confirmed(query_.to_header(test::bogus_block12.hash()), true));
@@ -253,14 +253,14 @@ BOOST_AUTO_TEST_CASE(electrum__blockchain_address_subscribe__progressive_notify_
     const auto notification2 = receive();
     REQUIRE_NO_THROW_TRUE(notification2.at("method").is_string());
     REQUIRE_NO_THROW_TRUE(notification2.at("params").is_array());
-    BOOST_CHECK_EQUAL(notification2.at("method").as_string(), "blockchain.address.subscribe");
+    BOOST_REQUIRE_EQUAL(notification2.at("method").as_string(), "blockchain.address.subscribe");
 
     const auto& params2 = notification2.at("params").as_array();
     BOOST_REQUIRE_EQUAL(params2.size(), 2u);
-    BOOST_CHECK(params2.at(0).is_string());
-    BOOST_CHECK(params2.at(1).is_string());
-    BOOST_CHECK_EQUAL(params2.at(0).as_string(), found_scripthash);
-    BOOST_CHECK_EQUAL(params2.at(1).as_string(), expected_confirm12);
+    BOOST_REQUIRE(params2.at(0).is_string());
+    BOOST_REQUIRE(params2.at(1).is_string());
+    BOOST_REQUIRE_EQUAL(params2.at(0).as_string(), found_scripthash);
+    BOOST_REQUIRE_EQUAL(params2.at(1).as_string(), expected_confirm12);
 }
 
 // blockchain.scripthash.subscribe
@@ -339,7 +339,7 @@ BOOST_AUTO_TEST_CASE(electrum__blockchain_scripthash_subscribe__repeat_call__ide
 {
     BOOST_REQUIRE(handshake(electrum::version::v1_1));
 
-    // This validates curosr/midstate consistency.
+    // This validates cursor/midstate consistency.
     BOOST_REQUIRE(query_.set(test::bogus_block10, database::context{ 0, 10, 0 }, false, false));
     BOOST_REQUIRE(query_.set(test::bogus_block11, database::context{ 0, 11, 0 }, false, false));
     BOOST_REQUIRE(query_.set(test::bogus_block12, database::context{ 0, 12, 0 }, false, false));
@@ -369,7 +369,7 @@ BOOST_AUTO_TEST_CASE(electrum__blockchain_scripthash_subscribe__progressive__exp
 {
     BOOST_REQUIRE(handshake(electrum::version::v1_1));
 
-    // This validates curosr/midstate consistency.
+    // This validates cursor/midstate consistency.
     BOOST_REQUIRE(query_.set(test::bogus_block10, database::context{ 0, 10, 0 }, false, false));
     BOOST_REQUIRE(query_.set(test::bogus_block11, database::context{ 0, 11, 0 }, false, false));
     BOOST_REQUIRE(query_.set(test::bogus_block12, database::context{ 0, 12, 0 }, false, false));
@@ -422,7 +422,7 @@ BOOST_AUTO_TEST_CASE(electrum__blockchain_scripthash_subscribe__progressive_noti
 {
     BOOST_REQUIRE(handshake(electrum::version::v1_1));
 
-    // This validates curosr/midstate consistency.
+    // This validates cursor/midstate consistency.
     BOOST_REQUIRE(query_.set(test::bogus_block10, database::context{ 0, 10, 0 }, false, false));
     BOOST_REQUIRE(query_.set(test::bogus_block11, database::context{ 0, 11, 0 }, false, false));
     BOOST_REQUIRE(query_.set(test::bogus_block12, database::context{ 0, 12, 0 }, false, false));
@@ -459,14 +459,14 @@ BOOST_AUTO_TEST_CASE(electrum__blockchain_scripthash_subscribe__progressive_noti
     const auto notification1 = receive();
     REQUIRE_NO_THROW_TRUE(notification1.at("method").is_string());
     REQUIRE_NO_THROW_TRUE(notification1.at("params").is_array());
-    BOOST_CHECK_EQUAL(notification1.at("method").as_string(), "blockchain.scripthash.subscribe");
+    BOOST_REQUIRE_EQUAL(notification1.at("method").as_string(), "blockchain.scripthash.subscribe");
 
     const auto& params1 = notification1.at("params").as_array();
     BOOST_REQUIRE_EQUAL(params1.size(), 2u);
-    BOOST_CHECK(params1.at(0).is_string());
-    BOOST_CHECK(params1.at(1).is_string());
-    BOOST_CHECK_EQUAL(params1.at(0).as_string(), found_scripthash);
-    BOOST_CHECK_EQUAL(params1.at(1).as_string(), expected_confirm11);
+    BOOST_REQUIRE(params1.at(0).is_string());
+    BOOST_REQUIRE(params1.at(1).is_string());
+    BOOST_REQUIRE_EQUAL(params1.at(0).as_string(), found_scripthash);
+    BOOST_REQUIRE_EQUAL(params1.at(1).as_string(), expected_confirm11);
 
     // Confirming block 12 only makes block 12 confirmed.
     BOOST_REQUIRE(query_.push_confirmed(query_.to_header(test::bogus_block12.hash()), true));
@@ -483,16 +483,15 @@ BOOST_AUTO_TEST_CASE(electrum__blockchain_scripthash_subscribe__progressive_noti
     const auto notification2 = receive();
     REQUIRE_NO_THROW_TRUE(notification2.at("method").is_string());
     REQUIRE_NO_THROW_TRUE(notification2.at("params").is_array());
-    BOOST_CHECK_EQUAL(notification2.at("method").as_string(), "blockchain.scripthash.subscribe");
+    BOOST_REQUIRE_EQUAL(notification2.at("method").as_string(), "blockchain.scripthash.subscribe");
 
     const auto& params2 = notification2.at("params").as_array();
     BOOST_REQUIRE_EQUAL(params2.size(), 2u);
-    BOOST_CHECK(params2.at(0).is_string());
-    BOOST_CHECK(params2.at(1).is_string());
-    BOOST_CHECK_EQUAL(params2.at(0).as_string(), found_scripthash);
-    BOOST_CHECK_EQUAL(params2.at(1).as_string(), expected_confirm12);
+    BOOST_REQUIRE(params2.at(0).is_string());
+    BOOST_REQUIRE(params2.at(1).is_string());
+    BOOST_REQUIRE_EQUAL(params2.at(0).as_string(), found_scripthash);
+    BOOST_REQUIRE_EQUAL(params2.at(1).as_string(), expected_confirm12);
 }
-
 
 // blockchain.scripthash.unsubscribe
 
@@ -606,7 +605,7 @@ BOOST_AUTO_TEST_CASE(electrum__blockchain_scriptpubkey_subscribe__repeat_call__i
 {
     BOOST_REQUIRE(handshake(electrum::version::v1_7));
 
-    // This validates curosr/midstate consistency.
+    // This validates cursor/midstate consistency.
     BOOST_REQUIRE(query_.set(test::bogus_block10, database::context{ 0, 10, 0 }, false, false));
     BOOST_REQUIRE(query_.set(test::bogus_block11, database::context{ 0, 11, 0 }, false, false));
     BOOST_REQUIRE(query_.set(test::bogus_block12, database::context{ 0, 12, 0 }, false, false));
@@ -636,7 +635,7 @@ BOOST_AUTO_TEST_CASE(electrum__blockchain_scriptpubkey_subscribe__progressive__e
 {
     BOOST_REQUIRE(handshake(electrum::version::v1_7));
 
-    // This validates curosr/midstate consistency.
+    // This validates cursor/midstate consistency.
     BOOST_REQUIRE(query_.set(test::bogus_block10, database::context{ 0, 10, 0 }, false, false));
     BOOST_REQUIRE(query_.set(test::bogus_block11, database::context{ 0, 11, 0 }, false, false));
     BOOST_REQUIRE(query_.set(test::bogus_block12, database::context{ 0, 12, 0 }, false, false));
@@ -689,7 +688,7 @@ BOOST_AUTO_TEST_CASE(electrum__blockchain_scriptpubkey_subscribe__progressive_no
 {
     BOOST_REQUIRE(handshake(electrum::version::v1_7));
 
-    // This validates curosr/midstate consistency.
+    // This validates cursor/midstate consistency.
     BOOST_REQUIRE(query_.set(test::bogus_block10, database::context{ 0, 10, 0 }, false, false));
     BOOST_REQUIRE(query_.set(test::bogus_block11, database::context{ 0, 11, 0 }, false, false));
     BOOST_REQUIRE(query_.set(test::bogus_block12, database::context{ 0, 12, 0 }, false, false));
@@ -726,14 +725,14 @@ BOOST_AUTO_TEST_CASE(electrum__blockchain_scriptpubkey_subscribe__progressive_no
     const auto notification1 = receive();
     REQUIRE_NO_THROW_TRUE(notification1.at("method").is_string());
     REQUIRE_NO_THROW_TRUE(notification1.at("params").is_array());
-    BOOST_CHECK_EQUAL(notification1.at("method").as_string(), "blockchain.scriptpubkey.subscribe");
+    BOOST_REQUIRE_EQUAL(notification1.at("method").as_string(), "blockchain.scriptpubkey.subscribe");
 
     const auto& params1 = notification1.at("params").as_array();
     BOOST_REQUIRE_EQUAL(params1.size(), 2u);
-    BOOST_CHECK(params1.at(0).is_string());
-    BOOST_CHECK(params1.at(1).is_string());
-    BOOST_CHECK_EQUAL(params1.at(0).as_string(), found_scripthash);
-    BOOST_CHECK_EQUAL(params1.at(1).as_string(), expected_confirm11);
+    BOOST_REQUIRE(params1.at(0).is_string());
+    BOOST_REQUIRE(params1.at(1).is_string());
+    BOOST_REQUIRE_EQUAL(params1.at(0).as_string(), found_scripthash);
+    BOOST_REQUIRE_EQUAL(params1.at(1).as_string(), expected_confirm11);
 
     // Confirming block 12 only makes block 12 confirmed.
     BOOST_REQUIRE(query_.push_confirmed(query_.to_header(test::bogus_block12.hash()), true));
@@ -750,14 +749,14 @@ BOOST_AUTO_TEST_CASE(electrum__blockchain_scriptpubkey_subscribe__progressive_no
     const auto notification2 = receive();
     REQUIRE_NO_THROW_TRUE(notification2.at("method").is_string());
     REQUIRE_NO_THROW_TRUE(notification2.at("params").is_array());
-    BOOST_CHECK_EQUAL(notification2.at("method").as_string(), "blockchain.scriptpubkey.subscribe");
+    BOOST_REQUIRE_EQUAL(notification2.at("method").as_string(), "blockchain.scriptpubkey.subscribe");
 
     const auto& params2 = notification2.at("params").as_array();
     BOOST_REQUIRE_EQUAL(params2.size(), 2u);
-    BOOST_CHECK(params2.at(0).is_string());
-    BOOST_CHECK(params2.at(1).is_string());
-    BOOST_CHECK_EQUAL(params2.at(0).as_string(), found_scripthash);
-    BOOST_CHECK_EQUAL(params2.at(1).as_string(), expected_confirm12);
+    BOOST_REQUIRE(params2.at(0).is_string());
+    BOOST_REQUIRE(params2.at(1).is_string());
+    BOOST_REQUIRE_EQUAL(params2.at(0).as_string(), found_scripthash);
+    BOOST_REQUIRE_EQUAL(params2.at(1).as_string(), expected_confirm12);
 }
 
 // blockchain.scriptpubkey.unsubscribe
